Update URL
Just happened upon this because I was trying to figure out which GH project my `CNAME` was pointing. Anyway, I see you linked to the blog post that I have been working to get up. I had not really finalized it yet, but I am super happy to hear you found it useful. I scrapped the VuePress site though, so the URL needs to be updated.